+++ packages/cloud-cognitive/src/components/Tearsheet/TearsheetShell.js
@@ -354,7 +354,13 @@ TearsheetShell.propTypes = {
     // one should ensure appropriate validation is done.
     PropTypes.shape({
       ...Button.propTypes,
-      kind: PropTypes.oneOf(['ghost', 'secondary', 'primary']),
+      kind: PropTypes.oneOf([
+        'ghost',
+        'danger--ghost',
+        'secondary',
+        'danger',
+        'primary',
+      ]),
       label: PropTypes.string,
       loading: PropTypes.bool,
       // we duplicate this Button prop to improve the DocGen here